test(demo): move test functions to own file
